Use ptcglb extension for asset import
PTC HACK, set this GLTFImporter to target .ptcglb instead of .glb so
that it does not conflict with MRTK. This prevents the errors from
showing up in the console when opening the project. It retains the
ability for PTC to use this importer if desired by using the .ptcglb
extension. Note that today we do not import GLBs as assets into a
Unity Project, we load them.

This commit should *not* be mainlined into the official Khronos
UnityGLTF.